FAFB v4: replace 2021-08-21 with 2021-12-08
These changes are in response to a variety of re- and mis-labelings
discovered in the previous groundtruth, including versus v3.

Broadly, this splits the data into several sets:

- A set of the highest confident labelings.
- A set of the v3 skeletons with a few corrections.
- New CATMAID and Flywire labelings for v4.
- A set with the corrected v3 labelings and the Kenyon cells (KCs) from
  v4.

To allow comparison between different sets, the confident set is used as
a "holdout" and withheld from inclusion in the other sets, except for a
special production split `skeleton_no_test_including_holdout` where it
is appended to the training data.